% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_belhadj2015.R
\name{belhadj2015}
\alias{belhadj2015}
\title{Fuzzy monetary poverty estimation}
\usage{
belhadj2015(x, z1, z2, z, b)
}
\arguments{
\item{x}{A numeric vector of a monetary variable (or poverty predicate)}

\item{z1}{Parameter}

\item{z2}{Parameter}

\item{z}{Parameter}

\item{b}{Parameter}
}
\value{
The fuzzy membership function as of Belhadj (2015).
}
\description{
Fuzzy monetary poverty estimation
}
