% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_ZBM.R
\name{bootP}
\alias{bootP}
\title{Fuzzy monetary poverty estimation}
\usage{
bootP(x, R = 500)
}
\arguments{
\item{x}{A numeric vector of a predicate variable (or poverty predicate)}

\item{R}{The number of bootstrap replicates (dafaults to 500)}
}
\value{
A matrix of bootrstapped percentiles
}
\description{
Calculates bootstrap percentiles from Zedini and Belhadj (2015)
}
\references{
Zedini, A., & Belhadj, B. (2015). A New Approach to Unidimensional Poverty Analysis: Application to the T unisian Case. Review of Income and Wealth, 61(3), 465-476.
}
