% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_belhadj2011.R
\name{fm_belhadj2011}
\alias{fm_belhadj2011}
\title{Fuzzy monetary poverty estimation}
\usage{
fm_belhadj2011(x, z_min, z_max, weight, breakdown, ID)
}
\arguments{
\item{x}{poverty predicate}

\item{z_min}{parameter}

\item{z_max}{parameter}

\item{weight}{A numeric vector of sampling weights. if NULL simple random sampling weights will be used.}

\item{breakdown}{A factor of sub-domains to calculate estimates for.}

\item{ID}{A numeric or character vector of IDs. if NULL (the default) it is set as the row sequence.}
}
\value{
a list containing the membership function values and its expected value
}
\description{
constructs fuzzy monetary poverty estimates as of Belhadj (2011)
}
