% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_belhadj2015.R
\name{fm_belhadj2015}
\alias{fm_belhadj2015}
\title{Fuzzy monetary poverty estimation}
\usage{
fm_belhadj2015(x, z1, z2, b, breakdown, weight)
}
\arguments{
\item{x}{A numeric vector of a monetary variable (or poverty predicate)}

\item{z1}{Parameter}

\item{z2}{Parameter}

\item{b}{Parameter}

\item{breakdown}{A factor of sub-domains to calculate estimates for (using the same alpha).}

\item{weight}{A numeric vector of sampling weights}
}
\value{
a list containing the fuzzy membership function and the value of z found as of Belhadj(2015).
}
\description{
Fuzzy monetary poverty estimation
}
