% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_equate.R
\name{fm_equate}
\alias{fm_equate}
\title{Fuzzy monetary poverty estimation}
\usage{
fm_equate(predicate.ord, weight.ord, interval, verbose)
}
\arguments{
\item{predicate.ord}{a sorted vector of a predicate variable}

\item{weight.ord}{a sorted vector of weights (in the same order of predicate.ord)}

\item{interval}{The interval to look for the solution of the equation.}

\item{verbose}{Logical. whether to print the proceeding of the procedure.}
}
\value{
the obtained exponent
}
\description{
Solves the non-linear equation in Betti et. al, 2018.
}
\details{
Calculates the exponent parameter alpha of the non-linear equation of Betti et al, 2018 so that
the expected value of the fuzzy membership function equated the head count ratio.
}
