% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_FL.R
\name{fm_FL}
\alias{fm_FL}
\title{Fuzzy monetary poverty estimation}
\usage{
fm_FL(predicate.ord, weight.ord)
}
\arguments{
\item{predicate.ord}{A sorted vector of a predicate variable (in ascending order).}

\item{weight.ord}{A sorted vector of weights (in the same order of s.ord)}
}
\value{
A numeric vector containing the estimated membership function.
}
\description{
This function calculates the fuzzy membership function as defined in Betti et. al, 2018.
}
