% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_verma2.R
\name{fm_mu2}
\alias{fm_mu2}
\title{Fuzzy predicate poverty estimation}
\usage{
fm_mu2(predicate.ord, weight.ord, alpha)
}
\arguments{
\item{predicate.ord}{A ordered numeric vector of a predicate variable (i.e. equivalised income or expenditure)}

\item{weight.ord}{A numeric vector of sampling weights. if NULL simple random sampling weights will be used.}

\item{alpha}{The value of the exponent in equation $E(mu)^(alpha-1) = HCR$. If NULL it is calculated so that it equates the expectation of the membership function to HCR}
}
\value{
A numeric vector containing the estimated membership function.
}
\description{
Fuzzy predicate poverty estimation
}
