% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_tfr.R
\name{fm_mu_TFR}
\alias{fm_mu_TFR}
\title{Fuzzy monetary poverty estimation}
\usage{
fm_mu_TFR(monetary.ord, weight.ord, alpha)
}
\arguments{
\item{monetary.ord}{A ordered numeric vector of a monetary variable (i.e. equivalised income or expenditure)}

\item{weight.ord}{A numeric vector of sampling weights. if NULL simple random sampling weights will be used.}

\item{alpha}{The value of the exponent in equation $E(mu)^(alpha-1) = HCR$. If NULL it is calculated so that it equates the expectation of the membership function to HCR}
}
\value{
A numeric vector containing the estimated membership function.
}
\description{
Fuzzy monetary poverty estimation
}
