% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_objective.R
\name{fm_objective}
\alias{fm_objective}
\title{Fuzzy monetary poverty estimation.}
\usage{
fm_objective(predicate.ord, weight.ord, alpha, HCR, fm, verbose)
}
\arguments{
\item{predicate.ord}{A sorted vector of a predicate variable (in ascending order).}

\item{weight.ord}{A sorted vector of weights (in the same order of s.ord)}

\item{alpha}{The value of the exponent parameter to use in the non-linear equation as of Betti et. al, 2018.}

\item{HCR}{The head count ratio.}

\item{fm}{the type of membership function to use}

\item{verbose}{prints the proceeding of the routine.}
}
\value{
The value of the objective function
}
\description{
Fuzzy monetary poverty estimation.
}
