% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_var.R
\name{fm_var}
\alias{fm_var}
\title{Fuzzy monetary poverty estimation}
\usage{
fm_var(
  income,
  weight,
  ID = NULL,
  HCR,
  breakdown = NULL,
  interval = c(1, 10),
  alpha = NULL,
  type = "bootstrap",
  R = 100,
  M = NULL,
  verbose = TRUE,
  stratum,
  psu,
  f = 0.01,
  ...
)
}
\arguments{
\item{income}{A numeric vector of incomes.}

\item{weight}{A numeric vector of sampling weights. if NULL simple random sampling weights will be used.}

\item{ID}{A numeric or character vector of IDs. if NULL (the default) it is set as the row sequence.}

\item{HCR}{The head count ratio.}

\item{breakdown}{A factor of sub-domains to calculate estimates for (using the same alpha).}

\item{interval}{The interval to look for the solution of the equation.}

\item{alpha}{The value of the exponent parameter to use in the non-linear equation as of Betti et. al, 2018.}

\item{type}{The variance estimation method chosen. One between `bootstrap` (default) or `jackknife`.}

\item{R}{The number of bootstrap replicates.}

\item{M}{The size of bootstrap samples. if NULL (default) is chosen as the sample size (i.e. the length of income)}

\item{verbose}{logical. whether to print the proceeding of the variance estimation procedure.}

\item{psu}{the vector identifying the psu (if 'jacknife' is chosen as variance estimation technique).}

\item{f}{the finite population correction fraction (if 'jacknife' is chosen as variance estimation technique).}

\item{...}{additional parameters}

\item{stratum.}{the vector identifying the stratum (if 'jacknife' is chosen as variance estimation technique).}
}
\value{
The estimate of variance with the method selected. if breakdown is not NULL, the variance is estimated for each sub-domain.
}
\description{
This function estimates the variance of the fuzzy monetary poverty index
}
