% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_ZBM.R
\name{fm_ZBM}
\alias{fm_ZBM}
\title{Fuzzy monetary poverty estimation}
\usage{
fm_ZBM(predicate, hh.size, weight, breakdown, k)
}
\arguments{
\item{predicate}{A numeric vector of a predicate variable (or poverty predicate)}

\item{hh.size}{A numeric vector with the size of the houshold}

\item{weight}{A numeric vector of sampling weights. if NULL simple random sampling weights will be used}

\item{breakdown}{A factor of sub-domains to calculate estimates for (using the same alpha).}

\item{k}{The number of change points locations to estimate}
}
\value{
The membership grades for each poverty state
}
\description{
Fuzzy monetary poverty estimation
}
\references{
Zedini, A., & Belhadj, B. (2015). A New Approach to Unidimensional Poverty Analysis: Application to the Tunisian Case. Review of Income and Wealth, 61(3), 465-476.
Belhadj, B., & Matoussi, M. S. (2010). Poverty in tunisia: A fuzzy measurement approach. Swiss Journal of Economics and Statistics, 146(2), 431-450.
}
