% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_ZBM.R
\name{FN}
\alias{FN}
\title{Fuzzy monetary poverty estimation}
\usage{
FN(x, a, b, c)
}
\arguments{
\item{x}{A numeric vector of a predicate variable (or poverty predicate)}

\item{a}{Membership function parameter}

\item{b}{Membership function parameter}

\item{c}{Membership function parameter}
}
\value{
The membrship grade
}
\description{
Fuzzy monetary poverty estimation
}
\references{
Zedini, A., & Belhadj, B. (2015). A New Approach to Unidimensional Poverty Analysis: Application to the T unisian Case. Review of Income and Wealth, 61(3), 465-476.
}
