% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_construct_all.R
\name{fs_construct_all}
\alias{fs_construct_all}
\title{Fuzzy supplementary poverty estimation.}
\usage{
fs_construct_all(
  data,
  weight = NULL,
  ID = NULL,
  dimensions,
  rho = NULL,
  HCR,
  interval = c(1, 10),
  alpha = NULL,
  breakdown = NULL
)
}
\arguments{
\item{data}{A matrix or a data frame of identified items (see Step 1 of Betti et. al, 2018)}

\item{weight}{A numeric vector of sampling weights. if NULL simple random sampling weights will be used}

\item{ID}{A numeric or character vector of IDs. if NULL (the default) it is set as the row sequence.}

\item{dimensions}{A numeric vector (of length  \code{ncol(data)}) of assignments of items in data to dimensions.}

\item{rho}{The critical value to be used for calculation of weights in the kendall correlation matrix.}

\item{HCR}{The value of the head count ratio.}

\item{interval}{A numeric vector of length two to look for the value of alpha (if not supplied).}

\item{alpha}{The value of the exponent in equation $E(mu)^(alpha-1) = HCR$. If NULL it is calculated so that it equates the expectation of the membership function to HCR.}

\item{breakdown}{A Dimension of sub-domains to calculate estimates for (using the same alpha). If numeric will be coerced to a Dimension.}
}
\value{
An object of class FuzzySupplementary containing the fuzzy membership function for each unit, the point estimate (i.e. the expected value of the function), and the alpha parameter.
}
\description{
Fuzzy supplementary poverty estimation.
}
\examples{
data("eusilc")
FS <- fs_construct_all(data = eusilc[,4:23], weight = eusilc$DB090, # step 2
                       dimensions = c(1,1,1,1,2,2,2,2,2,3,3,3,3,4,4,4,4,5,5,5), # step 3
                       rho = NULL, # steps 4 and 5
                       HCR = .12, # step 6
                       breakdown = eusilc$db040) # step 7 with breakdowns
summary(FS)
plot(FS)

}
