% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_mu.R
\name{fs_mu}
\alias{fs_mu}
\title{Fuzzy supplementary poverty estimation.}
\usage{
fs_mu(s.ord, w.ord, alpha)
}
\arguments{
\item{s.ord}{A vector of ordered deprivation scores.}

\item{w.ord}{A vector of ordered sampling weights. In the same order of \code{s.ord}.}

\item{alpha}{The alpha parameter.}
}
\value{
the fuzzy membership function.
}
\description{
Fuzzy supplementary poverty estimation.
}
