% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_var.R
\name{fs_var}
\alias{fs_var}
\title{Fuzzy supplementary poverty estimation.}
\usage{
fs_var(
  data,
  weight = NULL,
  ID = NULL,
  dimensions,
  breakdown = NULL,
  HCR,
  alpha,
  rho = NULL,
  type = "bootstrap",
  M = NULL,
  R = 2,
  verbose = TRUE,
  stratum = NULL,
  psu = NULL,
  f = 0.01
)
}
\arguments{
\item{data}{data a numeric matrix or data frame (? davvero?) of items.}

\item{weight}{a numeric vector of sampling weights. if NULL simple random sampling weights will be used}

\item{ID}{a numeric or character vector of IDs.}

\item{dimensions}{a numeric vector (of length  `ncol(data)`) of assignments of items in data to dimensions.}

\item{breakdown}{a factor. if supplied estimates will be calculated for each level (using the same alpha).}

\item{HCR}{The value of the head count ratio.}

\item{alpha}{The value of the exponent in equation $E(mu)^(\alpha-1) = HCR$. If NULL it is calculated so that it equates the expectation of the membership function to HCR.}

\item{rho}{The critical value of the correlation coefficient.}

\item{type}{The variance estimation method, default is bootstrap.}

\item{M}{The size of bootstrap samples. Default is `nrow(data)`.}

\item{R}{The number of bootstrap replicates}

\item{verbose}{logical. whether to print the proceeding of the variance estimation procedure.}

\item{stratum}{the vector identifying the stratum (if 'jacknife' is chosen as variance estimation technique).}

\item{psu}{the vector identifying the psu (if 'jacknife' is chosen as variance estimation technique).}

\item{f}{the finite population correction fraction (if 'jacknife' is chosen as variance estimation technique).}
}
\value{
A list containing the estimated variance.
}
\description{
Fuzzy supplementary poverty estimation.
}
