% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_ZBM.R
\name{Fuzzy_conv}
\alias{Fuzzy_conv}
\title{Fuzzy monetary poverty estimation}
\usage{
Fuzzy_conv(x)
}
\arguments{
\item{x}{A numeric vector of fuzzy numbers}
}
\value{
the convolution of fuzzy numbers
}
\description{
Fuzzy monetary poverty estimation
}
