% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyPovertyR-package.R
\docType{package}
\name{FuzzyPovertyR-package}
\alias{FuzzyPovertyR}
\alias{FuzzyPovertyR-package}
\title{FuzzyPovertyR: Estimation of Fuzzy Poverty Measures}
\description{
Estimates fuzzy measures of poverty and deprivation. It also estimates the sampling variance of these measures using bootstrap or jackknife repeated replications.
}
\author{
\strong{Maintainer}: Federico Crescenzi \email{federico.crescenzi@unitus.it} (\href{https://orcid.org/0000-0002-0701-4398}{ORCID})

Authors:
\itemize{
  \item Lorenzo Mori \email{lorenzo.mori7@unibo.it} (\href{https://orcid.org/0000-0003-4152-9712}{ORCID})
}

Other contributors:
\itemize{
  \item Gianni Betti \email{betti@unisi.it} (\href{https://orcid.org/0000-0002-0414-7806}{ORCID}) [contributor]
}

}
\keyword{internal}
