% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyPoverty_class.R
\name{FuzzySupplementary}
\alias{FuzzySupplementary}
\title{s3 class fuzzy poverty}
\usage{
FuzzySupplementary(x)
}
\arguments{
\item{x}{an object}
}
\value{
an object of class FuzzyPoverty
}
\description{
s3 class fuzzy poverty
}
