% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{HCR}
\alias{HCR}
\title{Head Count Ratio (HCR)}
\usage{
HCR(monetary, weight = NULL, p = 0.5, q = 0.6, poverty.line = NULL)
}
\arguments{
\item{monetary}{A numeric vector of a monetary variable (i.e. income or expenditure)}

\item{weight}{A numeric vector of sampling weights. if NULL simple random sampling weights will be used.}

\item{p}{The quantile to be calculated from the monetary variable. Default is the median.}

\item{q}{The percentage of the quantile to be used in determining the poverty line. default is 0.6.}

\item{poverty.line}{The poverty line. If it is NULL it is estimated from data.}
}
\value{
A list containing the classification of the units as poor (TRUE) and not-poor (FALSE), the estimated Head Count Ratio, and the poverty line.
}
\description{
This function calculates the head count ratio.
}
\details{
The head count ration is defined as the sum of the sampling weight of statistical units
whose vale of the monetary variable is below the poverty line. The poverty line is usually defined as a
fraction of a weighted quantile (in official statistics the median) of the monetary distribution
}
\examples{
N <- 100
p <- 0.5
q <- 0.6
monetary <- rchisq(N, 15) # monetary variable
HCR(monetary)

}
