% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_ZBM.R
\name{MemberhsipGradesMatrix}
\alias{MemberhsipGradesMatrix}
\title{Fuzzy monetary poverty estimation}
\usage{
MemberhsipGradesMatrix(x, P)
}
\arguments{
\item{x}{A numeric vector of a predicate variable (or poverty predicate)}

\item{P}{A matrix of bootstapped percentiles}
}
\value{
The membership grade matrix
}
\description{
Fuzzy monetary poverty estimation
}
