% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_method.R
\name{plot.FuzzyMonetary}
\alias{plot.FuzzyMonetary}
\title{The plot of a FuzzyMonetary object}
\usage{
\method{plot}{FuzzyMonetary}(x, ...)
}
\arguments{
\item{x}{An object of class "FuzzyMonetary"}

\item{...}{Additional options}
}
\value{
The plot
}
\description{
plot method for class "FuzzyMonetary"
}
