% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_plot.R
\name{plot.FuzzySupplementary}
\alias{plot.FuzzySupplementary}
\title{The plot of a FuzzySupplementary object}
\usage{
\method{plot}{FuzzySupplementary}(x, ...)
}
\arguments{
\item{x}{An object of class "FuzzySupplementary"}

\item{...}{Additional options}
}
\value{
The plot
}
\description{
plot method for class "FuzzySupplementary"
}
