% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_method.R
\name{summary.FuzzyMonetary}
\alias{summary.FuzzyMonetary}
\title{The summary of a FuzzyMonetary object}
\usage{
\method{summary}{FuzzyMonetary}(object, ...)
}
\arguments{
\item{object}{An object of class "FuzzyMonetary"}

\item{...}{Additional options}
}
\value{
The summary method for class "FuzzyMonetary"
}
\description{
Summary method for class "FuzzyMonetary"
}
