% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_method.R
\name{summary.FuzzyPoverty}
\alias{summary.FuzzyPoverty}
\title{The summary of a FuzzyPoverty object}
\usage{
\method{summary}{FuzzyPoverty}(obj)
}
\arguments{
\item{An}{object of class "FuzzyPoverty"}
}
\value{
The summary method for class "FuzzyPoverty"
}
\description{
Summary method for class "FuzzyPoverty"
}
