% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_summary.R
\name{summary.FuzzySupplementary}
\alias{summary.FuzzySupplementary}
\title{The summary of a FuzzySupplementary object}
\usage{
\method{summary}{FuzzySupplementary}(object, ...)
}
\arguments{
\item{object}{An object of class "FuzzySupplementary"}

\item{...}{Additional options}
}
\value{
The summary method for class "FuzzySupplementary"
}
\description{
Summary method for class "FuzzySupplementary"
}
