% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_wb.R
\name{wb.jh}
\alias{wb.jh}
\title{Fuzzy supplementary poverty estimation.}
\usage{
wb.jh(j, step2, dimensions, rho, ...)
}
\arguments{
\item{j}{The number identifying one of the latent dimensions discovered}

\item{step2}{The result obtained from Step2 of the fuzzy supplementary procedure.}

\item{dimensions}{A numeric vector enumerating the dimensions discovered}

\item{rho}{The critical value to be used for calculation of weights in the kendall correlation matrix.}

\item{...}{other parameters}
}
\value{
The weight obtained for item j in dimension h as of Formula 12 of Betti and Verma 2008.
}
\description{
Fuzzy supplementary poverty estimation.
}
\references{
Betti, G., & Verma, V. (2008). Fuzzy measures of the incidence of relative poverty and deprivation: a multi-dimensional perspective. Statistical Methods and Applications, 17(2), 225-250.
}
