% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_belhadj2015.R
\name{z_fun}
\alias{z_fun}
\title{Fuzzy monetary poverty estimation}
\usage{
z_fun(x, z1, z2, b)
}
\arguments{
\item{x}{A numeric vector of a monetary variable (or poverty predicate)}

\item{z1}{Parameter}

\item{z2}{Parameter}

\item{b}{Parameter}
}
\value{
the difference between memberhip functions
}
\description{
Fuzzy monetary poverty estimation
}
