% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_equate.R
\name{fm_equate}
\alias{fm_equate}
\title{Fuzzy monetary poverty estimation}
\usage{
fm_equate(monetary.ord, weight.ord, interval)
}
\arguments{
\item{monetary.ord}{a sorted vector of a monetary variable}

\item{weight.ord}{a sorted vector of weights (in the same order of monetary.ord)}

\item{interval}{The interval to look for the solution of the equation.}
}
\value{
the obtained exponent alfa
}
\description{
Solves the non-linear equation in Betti et. al, 2018.
}
\details{
Calculates the exponent parameter alpha of the non-linear equation of Betti et al, 2018 so that
the expected value of the fuzzy membership function equated the head count ratio.
}
