% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_var.R
\name{fm_var}
\alias{fm_var}
\title{Fuzzy monetary poverty estimation}
\usage{
fm_var(
  monetary,
  weight,
  ID = NULL,
  HCR,
  breakdown = NULL,
  interval = c(1, 10),
  alpha = NULL,
  type = "bootstrap",
  R = 100,
  M = NULL,
  stratum,
  psu,
  f = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{monetary}{A numeric vector of a monetary variable (i.e. equivalised income or expenditure)}

\item{weight}{A numeric vector of sampling weights. if NULL simple random sampling weights will be used.}

\item{ID}{A numeric or character vector of IDs. if NULL (the default) it is set as the row sequence.}

\item{HCR}{The head count ratio.}

\item{breakdown}{A factor of sub-domains to calculate estimates for (using the same alpha). Ff numeric will be coherced to a factor.}

\item{interval}{The interval to look for the solution of the equation.}

\item{alpha}{The value of the exponent in equation $E(mu)^(\alpha-1) = HCR$. If NULL it is calculated so that it equates the expectation of the membership function to HCR.}

\item{type}{The variance estimation method chosen. One between \code{bootstrap} (default) or \code{jackknife}.}

\item{R}{The number of bootstrap replicates. Default is 500.}

\item{M}{The size of bootstrap samples. Default is \code{nrow(data)}.}

\item{stratum}{The vector identifying the stratum (if 'jackknife' is chosen as variance estimation technique).}

\item{psu}{The vector identifying the psu (if 'jacknife' is chosen as variance estimation technique).}

\item{f}{The finite population correction fraction (if 'jackknife' is chosen as variance estimation technique).}

\item{verbose}{Logical. whether to print the proceeding of the variance estimation procedure.}
}
\value{
The estimate of variance with the method selected. if breakdown is not NULL, the variance is estimated for each sub-domain.
}
\description{
This function estimates the variance of the fuzzy monetary poverty index
}
\examples{
data(eusilc)
HCR <- 0.14
fm_var(monetary = eusilc$red_eq, weight = eusilc$DB090, ID = eusilc$ID, HCR = HCR, type = 'bootstrap', alpha = 5, R = 100)
fm_var(monetary = eusilc$red_eq, weight = eusilc$DB090, ID = eusilc$ID, HCR = HCR, type = 'bootstrap', alpha = 5, R = 100, breakdown = eusilc$db040)
fm_var(monetary = eusilc$red_eq, weight = eusilc$DB090, ID = eusilc$ID, HCR = HCR, type = 'jackknife', alpha = 5, stratum = eusilc$stratum, psu = eusilc$psu)
fm_var(monetary = eusilc$red_eq, weight = eusilc$DB090, ID = eusilc$ID, HCR = HCR, type = 'jackknife', alpha = 5, stratum = eusilc$stratum, psu = eusilc$psu, breakdown = eusilc$db040)
}
