% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_ZBM.R
\name{fm_ZBM}
\alias{fm_ZBM}
\title{Fuzzy monetary poverty estimation}
\usage{
fm_ZBM(monetary, hh.size, weight, breakdown, k)
}
\arguments{
\item{monetary}{A numeric vector of a monetary variable (or poverty predicate)}

\item{hh.size}{A numeric vector with the size of the houshold}

\item{weight}{A numeric vector of sampling weights. if NULL simple random sampling weights will be used}

\item{k}{The number of change points locations to estimate}
}
\value{
The membership grades for each poverty state
}
\description{
Fuzzy monetary poverty estimation
}
\examples{
# three poverty states
x1 = rchisq(500, 15)
x2 = rchisq(500, 10)
x3 = rchisq(500, 30)
x = c(x1, x2, x3)
w = sample(1:5, 1500, replace = T) # household sizes
breakdown = sample(letters[1:3], size = 1500, replace = T,  prob = c(3,1,1))

P <- bootP(x)
MGM <- MemberhsipGradesMatrix(x, P)
fm_ZBM(monetary = x, hh.size = w, weight = NULL, k = 3, breakdown = breakdown)

}
\references{
Zedini, A., & Belhadj, B. (2015). A New Approach to Unidimensional Poverty Analysis: Application to the Tunisian Case. Review of Income and Wealth, 61(3), 465-476.
Belhadj, B., & Matoussi, M. S. (2010). Poverty in tunisia: A fuzzy measurement approach. Swiss Journal of Economics and Statistics, 146(2), 431-450.
}
