% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_construct.R
\name{fs_construct}
\alias{fs_construct}
\title{Fuzzy supplementary poverty estimation.}
\usage{
fs_construct(steps3_5, weight, alpha, breakdown = NULL, ...)
}
\arguments{
\item{steps3_5}{The results from previous steps.}

\item{weight}{A vector of sampling weights. If it is NULL (the default) weights are assigned assuming simple random sampling of units.}

\item{alpha}{The alfa parameter.}

\item{breakdown}{A factor of sub-domains to calculate estimates for (using the same alpha).}

\item{...}{Other parameters.}
}
\value{
A list of results containing the fuzzy meambership function for each unit, the point estimate (i.e. the expected value of the function), and the alpha parameter.
}
\description{
Constructs the fuzzy supplementary poverty measure based on Steps1-6.
}
