% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_equate.R
\name{fs_equate}
\alias{fs_equate}
\title{Fuzzy supplementary poverty estimation}
\usage{
fs_equate(steps3_5, weight, HCR, interval = c(1, 10))
}
\arguments{
\item{steps3_5}{The results obtained from \code{fs_weight}.}

\item{weight}{A vector of sampling weights. If it is NULL (the default) weights are assigned assuming simple random sampling of units.}

\item{HCR}{The head count ratio.}

\item{interval}{The range to look for the value of alpha.}
}
\value{
The alpha parameter that solves the non-linear equation $E\link{mu} = HCR$
}
\description{
This function solves $E(mu)^(\alpha-1) = HCR$ for alpha.
}
\examples{
alpha <- fs_equate(steps3_5 = steps3_5, weight = eusilc$DB090, HCR = .16, interval = c(1,10))
}
