% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_weight.R
\name{fs_weight}
\alias{fs_weight}
\title{Fuzzy supplementary poverty estimation}
\usage{
fs_weight(dimensions, step2, rho = NULL, ...)
}
\arguments{
\item{dimensions}{A vector of assignments to a latent dimension for each column in data.}

\item{step2}{The data frame resulting from step2.}

\item{rho}{The critical value to be used for calculation of weights in the kendall correlation matrix.}

\item{...}{}
}
\value{
da vedere
}
\description{
Calculates the weights of dimensions discovered after factor analysis.
}
