% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_weight.R
\name{fs_weight}
\alias{fs_weight}
\title{Fuzzy supplementary poverty estimation}
\usage{
fs_weight(dimensions, step2, rho = NULL)
}
\arguments{
\item{dimensions}{A numeric vector (of length  \code{ncol(data)}) of assignments of items in data to dimensions.}

\item{step2}{The data frame resulting from step2.}

\item{rho}{The critical value to be used for calculation of weights in the kendall correlation matrix.}
}
\value{
A data frame of weights and deprivation scores in each dimension identified.
}
\description{
Calculates the weights of dimensions discovered after factor analysis.
}
