% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyPovertyR-package.R
\docType{package}
\name{FuzzyPovertyR-package}
\alias{FuzzyPovertyR}
\alias{FuzzyPovertyR-package}
\title{FuzzyPovertyR: Estimation of fuzzy poverty measures}
\description{
This package estimates fuzzy monetary and fuzzy supplementary poverty measures. It also estimates the sampling variance associated with the point estimates.
}
\author{
\strong{Maintainer}: Federico Crescenzi \email{federico.crescenzi@unitus.it} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Lorenzo Mori \email{lorenzo.mori7@unibo.it}
}

Other contributors:
\itemize{
  \item Gianni Betti \email{betti@unisi.it} [contributor]
}

}
\keyword{internal}
