% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_modified_sum.R
\name{modifiedSum}
\alias{modifiedSum}
\title{Fuzzy supplementary poverty estimation.}
\usage{
modifiedSum(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
the sum of x,y
}
\description{
Fuzzy supplementary poverty estimation.
}
