% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_method.R
\name{plot.FuzzyMonetary}
\alias{plot.FuzzyMonetary}
\title{The plot of a FuzzyMonetary object}
\usage{
\method{plot}{FuzzyMonetary}(x, ...)
}
\arguments{
\item{x}{An object of class "FuzzyMonetary"}

\item{...}{Additional options}
}
\value{
The plot
}
\description{
plot method for class "FuzzyMonetary"
}
\examples{

#The following example is based on the dataset eusilc
#included in the package.


#fm = "verma"

index = fm_construct(predicate = eusilc$eq_income, weight = eusilc$DB090,
             fm = "verma", HCR = 0.154, ID = eusilc$ID)

plot(index)
}
