% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_method.R
\name{plot.FuzzyPoverty}
\alias{plot.FuzzyPoverty}
\title{The plot of a FuzzyPoverty object}
\usage{
\method{plot}{FuzzyPoverty}(x, ...)
}
\arguments{
\item{x}{An object of class "FuzzyPoverty"}

\item{...}{Additional options}
}
\value{
The plot
}
\description{
plot method for class "FuzzyPoverty"
}
