% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_method.R
\name{summary.FuzzyPoverty}
\alias{summary.FuzzyPoverty}
\title{The summary of a FuzzyPoverty object}
\usage{
\method{summary}{FuzzyPoverty}(object, ...)
}
\arguments{
\item{object}{An object of class "FuzzyPoverty"}

\item{...}{Additional options}
}
\value{
The summary method for class "FuzzyPoverty"
}
\description{
Summary method for class "FuzzyPoverty"
}
