% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_belhadj2015.R
\name{ub1}
\alias{ub1}
\title{Fuzzy monetary poverty estimation}
\usage{
ub1(x, z1, b)
}
\arguments{
\item{x}{A numeric vector of a monetary variable (or poverty predicate)}

\item{z1}{Parameter}

\item{b}{Parameter}
}
\description{
Fuzzy monetary poverty estimation
}
