% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_belhadj2015.R
\name{ub2}
\alias{ub2}
\title{Fuzzy monetary poverty estimation}
\usage{
ub2(x, z2, b)
}
\arguments{
\item{x}{A numeric vector of a monetary variable (or poverty predicate)}

\item{z2}{Parameter}

\item{b}{Parameter}
}
\description{
Fuzzy monetary poverty estimation
}
