% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_quantile.R
\name{weighted_quantile}
\alias{weighted_quantile}
\title{Calculating a weighted quantile.}
\usage{
weighted_quantile(x, w, p)
}
\arguments{
\item{x}{a numeric vector representing a statistical variable}

\item{w}{a numeric vector containing sampling weights}

\item{p}{the order of quantile}
}
\value{
a weighted quantile
}
\description{
This function calculates a weighted quantile
}
\examples{

N <- 100
p <- 0.5
x <- rchisq(N, 15) # monetary variable
w <- rep(1/N, N) # sampling weights
weighted.quantile(x,w,p=0.5)
}
